making createResourceGroup private and moving readJson to deployment.go
